memory leaks fixed; splint errors corrected; read_tags function partially implemented;

+++ src/common.h
@@ -30,7 +30,7 @@ typedef struct {
   //< number of tasks in a multitask setting
   size_t m_n_tasks;
   //< maximum number of classes for each task
-  size_t m_max_classes[MAX_TASKS];
+  int m_max_classes[MAX_TASKS];
 } multiclass_t;
